# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: 小红书热点搭子 · 热点相关性 API
  version: 0.1.0
  summary: 输入热点，在本地素材库中检索可蹭素材并产出文案初稿
  description: |
    ## 本文件的地位

    本文件是 API 的**唯一事实源**。代码向契约对齐；改接口必须先改本文件，
    再改实现，并通过 CI 中的契约一致性校验（见 docs/backlog.md 的 S4.6）。

    ## 设计前提

    - **本地单人工具**：不提供鉴权、不区分用户、不做限流。此处写"不适用"而非留空，
      是为了让面试追问时能给出明确答案，而不是"还没来得及做"。
    - **不做运行时降级**：依赖不可用时返回错误码，不返回降级结果。
    - **分析是异步任务**：`POST /api/analyze` 立即返回 `job_id`，进度与结果分开查询。
    - **API 统一 `/api` 前缀**（见 ADR 0009）：根路径与其余非 `/api` 路径全部交给前端单页应用，
      前端路由与 API 路径互不干扰。
    - **前端不在本契约内**：单页应用的构建产物由服务端静态挂载，技术栈见 `docs/技术栈.md`。

    ## 通用约定

    - 字段名一律 snake_case；时间戳为 RFC 3339（UTC）。
    - 除报告接口外，响应体均为 `application/json`。
    - 请求可携带 `X-Request-ID`，响应同值回写；未携带时由服务端生成，并贯穿日志与调用追踪。
    - `topk` 默认 5，允许范围 1–20；`hotspots` 允许 1–10 条。

    ## 错误码

    | code | HTTP | 含义 |
    | --- | --- | --- |
    | `bad_request` | 400 | 请求格式或参数不合法 |
    | `validation_error` | 422 | 字段级校验失败（如 topk 越界） |
    | `not_found` | 404 | job_id 或 run_id 不存在 |
    | `conflict` | 409 | 当前状态不允许该操作（如运行尚未完成就取结果） |
    | `payload_too_large` | 413 | 请求体超过体积上限（如上传超过 `[upload].max_size_gb`） |
    | `dependency_unavailable` | 503 | 数据库或 Redis 不可用，或模型配置缺失 |
    | `upstream_error` | 502 | 上游模型返回错误或连续重试失败 |
    | `internal_error` | 500 | 未分类的内部错误 |
  license:
    name: Proprietary
servers:
  - url: http://localhost:8000
    description: 本地与演示环境。API 统一挂载在 /api 前缀下；根路径与前端路由由单页应用处理（见 ADR 0009）。
security: []
...
paths:
...
  /api/chat/sessions:
    post:
      tags:
        - 对话
      summary: 新建对话会话
      description: |
        建一个空会话（前端「新对话」）。标题留空，等**首条用户消息**落库时取前 20 字自动补上；
        会话与消息只在本机数据库里，附件落在 `runs/_chat/<session_id>/`。
      operationId: createChatSession
      parameters:
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatSessionCreate'
      responses:
        '201':
          description: 已创建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatSession'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/ValidationError'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
    get:
      tags:
        - 对话
      summary: 列出对话会话
      description: 按最近活跃（`updated_at`）倒序分页；每项带消息数与末条消息预览（超 60 字截断）。
      operationId: listChatSessions
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: 会话列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatSessionList'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
  /api/chat/sessions/{session_id}:
    get:
      tags:
        - 对话
      summary: 读取会话与消息
      description: |
        返回会话与它的消息（按时间升序）。`status=running` 的 assistant 消息代表"这一轮还没跑完"，
        前端刷新后用轮询补齐；超过 `[chat].turn_stale_seconds` 仍未收尾的会被标成 `interrupted`。
      operationId: getChatSession
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          required: false
          description: 只取最近 N 条消息（不传 = 全部）
          schema:
            type: integer
            minimum: 1
            maximum: 200
      responses:
        '200':
          description: 会话详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatSessionDetail'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
  /api/chat/sessions/{session_id}/messages:
    post:
      tags:
        - 对话
      summary: 发一条消息（SSE 流式）
      description: |
        一次 multipart 请求带上文字与（可选的）一张图片，返回 `text/event-stream`：
        `turn_started` → （可选）`tool_started` / `tool_progress` / `tool_finished` →
        `text_delta`（真 token 增量）→ `turn_finished`（带 run_id）；失败则发一条 `error`
        并结束这一轮。每一行 `data:` 是一个 JSON 对象。
        **断开浏览器只停转发**：这一轮照常跑完并落库（库里能看到 assistant 消息的终态）。
        图片只接受 jpg / png / webp（按 magic bytes），单张上限 `[chat].attachment_max_mb`。
      operationId: postChatMessage
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                text:
                  type: string
                  description: 用户这条消息的文字（与图片至少给一个）
                file:
                  type: string
                  format: binary
                  description: 可选的热点图片（jpg / png / webp）
      responses:
        '200':
          description: SSE 事件流
          content:
            text/event-stream:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
  /api/chat/sessions/{session_id}/attachments/{message_id}/{index}:
    get:
      tags:
        - 对话
      summary: 读取会话里的附件
      description: |
        按会话 + 消息 + 下标取回附件原图（回看历史会话时用）。消息不属于该会话、下标越界、
        文件缺失一律 404，不区分原因（与关键帧接口同一口径，避免暴露内部布局）。
      operationId: getChatAttachment
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: message_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: index
          in: path
          required: true
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          description: 附件图片
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
components:
  parameters:
    RequestId:
      name: X-Request-ID
      in: header
      required: false
      description: 客户端生成的请求标识；未提供时由服务端生成。用于贯穿日志、调用追踪与数据库记录。
      schema:
        type: string
        maxLength: 128
...
  responses:
    BadRequest:
      description: 请求格式或参数不合法
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: 字段级校验失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    DependencyUnavailable:
      description: 依赖不可用（不做降级，直接返回错误）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ChatSessionCreate:
      type: object
      description: 建会话请求体；可以不传（标题留空，等首条用户消息自动补）。
      additionalProperties: false
      properties:
        title:
          type: string
    ChatSession:
      type: object
      required:
        - session_id
        - title
        - created_at
        - updated_at
      properties:
        session_id:
          type: string
          format: uuid
        title:
          type: string
          description: 取首条用户消息前 20 字；未发言时为「新对话」，不可编辑
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ChatSessionSummary:
      type: object
      required:
        - session_id
        - title
        - created_at
        - updated_at
        - message_count
        - last_message_preview
      properties:
        session_id:
          type: string
          format: uuid
        title:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        message_count:
          type: integer
          minimum: 0
        last_message_preview:
          type: string
          description: 末条消息正文（超 60 字截断加省略号；没有消息时为空串）
    ChatSessionList:
      type: object
      required:
        - items
        - total
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ChatSessionSummary'
        total:
          type: integer
          minimum: 0
    ChatAttachment:
      type: object
      required:
        - kind
        - name
        - mime
        - bytes
        - url
      properties:
        kind:
          type: string
          enum:
            - image
        name:
          type: string
        mime:
          type: string
        bytes:
          type: integer
          minimum: 0
        url:
          type: string
          description: 取回附件的地址（可直接当 `<img src>`）
    ChatToolCall:
      type: object
      description: 一次工具调用的过程记录（挂在 assistant 消息上）。
      required:
        - tool
        - status
        - args
      properties:
        tool:
          type: string
        status:
          type: string
          enum:
            - succeeded
            - failed
        args:
          type: object
          additionalProperties: true
        run_id:
          oneOf:
            - type: string
              format: uuid
            - type: 'null'
        cost_cny:
          type: number
        latency_ms:
          type: integer
          minimum: 0
        error:
          type: string
        steps:
          type: array
          items:
            type: object
            additionalProperties: true
    ChatMessage:
      type: object
      required:
        - message_id
        - session_id
        - role
        - content
        - status
        - attachments
        - tool_calls
        - prompt_versions
        - created_at
      properties:
        message_id:
          type: string
          format: uuid
        session_id:
          type: string
          format: uuid
        role:
          type: string
          enum:
            - user
            - assistant
        content:
          type: string
        status:
          type: string
          enum:
            - running
            - succeeded
            - failed
            - interrupted
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/ChatAttachment'
        tool_calls:
          type: array
          items:
            $ref: '#/components/schemas/ChatToolCall'
        prompt_versions:
          type: object
          additionalProperties:
            type: integer
        run_id:
          oneOf:
            - type: string
              format: uuid
            - type: 'null'
        cost_cny:
          type: number
        latency_ms:
          type: integer
          minimum: 0
        error:
          oneOf:
            - type: string
            - type: 'null'
        created_at:
          type: string
          format: date-time
    ChatSessionDetail:
      type: object
      required:
        - session_id
        - title
        - created_at
        - updated_at
        - messages
      properties:
        session_id:
          type: string
          format: uuid
        title:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: '#/components/schemas/ChatMessage'
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          enum:
            - bad_request
            - validation_error
            - not_found
            - conflict
            - payload_too_large
            - dependency_unavailable
            - upstream_error
            - internal_error
        message:
          type: string
        detail:
          type: object
          additionalProperties: true